requirements.txt
fastapi
uvicorn
seleniumbase
pyautogui
pytesseract>=0.3.10
Pillow
python-dotenv
google-generativeai
